10/02/2018

+++ script/experiment/train_pcb.py
@@ -444,7 +444,7 @@ def main():
       print()
       print('criterion:', len([criterion(logits, labels_var) for logits in logits_list]))
       #loss = torch.sum(torch.cat([criterion(logits, labels_var) for logits in logits_list]))
-      loss = torch.sum([criterion(logits, labels_var) for logits in logits_list])
+      loss = torch.sum(torch.FloatTensor([criterion(logits, labels_var) for logits in logits_list]))
       optimizer.zero_grad()
       loss.backward()
       optimizer.step()
